Rename project to uvtool
We decided that "cloud" could be misleading. This tool does not provide
Ubuntu Cloud services like OpenStack and juju.

We decided to rename the tool to "uvtool" for now. This is intended to
mean "Ubuntu Virtualisation Tools", or hopefully "Universal
Virtualisation Tools" in the future.

We accept that this isn't a great name, but can't think of anything
better right now, and need to get the tool usable for juju in time for
Saucy's release. I am not averse to renaming this tool again later if
someone can suggest something better.

We decided on:

uvt-kvm for the core creation tool
uvt-simplestreams-libvirt to pull in Ubuntu Cloud images into a local
    libvirt pool.
uvtool for the libvirt pool name
uvtool for the Python namespace
uvtool for the source package in Ubuntu

An "apt-file find uvtool" found me nothing.

Google finds me nothing that looks likely to conflict.

launchpad.net/uvtools does exist, but is unrelated to virtualisation. I
don't think there will be any confusion here.

uvt-kvm is libvirt specific, but I think this is OK as our preferred
method of getting featureful kvm instances is through libvirt (eg.
our OpenStack offering does this too). This way we get awesome stuff
like automatic AppArmor-constrained instances.

If in the future we add a non-libvirt kvm tool, we could call it
uvt-kvmx or something.