Depend on cloud-image-utils >= 0.27
This is the release version in Saucy that contains cloud-localds. Older
releases did not contain cloud-localds.

We only need cloud-image-utils now, not cloud-utils, since the package
has been split into two and our needs lie with the "image" side of the
split, as opposed to the "guest" side.

This change allows a no-change backport to Precise, since cloud-localds
wasn't present in the Precise release, but will be available from the
cloud-tools cloud archive pocket and so we need to be explicit for
apt-get to pull it in.