Skip some tests when backported to Ubuntu Precise
We need a more recent version of mock, but don't have it on Ubuntu
Precise, to which we want uvtool to be backportable to for now.

Skipping tests shouldn't be so bad as in order for the backport process
to proceed, tests will have already passed against a newer version of
mock.