Default to 'ubuntu' login name for ssh
If the user hasn't specified the username via the -l option or via '@',
then default to the username 'ubuntu' instead of nothing at all, since
'ubuntu' is the default username on Ubuntu cloud images. Otherwise, ssh
uses the local username, which is unlikely to match when using uvt-kvm
from a typical desktop machine.

This does mean that uvt-kvm will override any default "User"
configuration option specified in the user's configuration file for ssh.
The user can always run "ssh" directly instead of through the "uvt-kvm
ssh" wrapper for full functionality, which seems reasonable if the user
has an ssh configuration for the host anyway. If there is any other way
to have uvt-kvm supply a default user that doesn't override an ssh
configuration file, then I would be happy to use that instead. In the
meantime, this is a trade-off; I think that on balance the more common
case of not having any configuration should work automatically and with
less intervention.

I accidentally included the default_login_name part of this feature in a
previous commit; it had no effect previously, but is now being used for
this feature.

LP: #1280588