Drop "uvt-kvm import"
This is a relic that was useful prior to simplestreams support. I don't
intend to support this subcommand as part of a stable API, since I don't
think anybody will need it in the common case. In particular, the uvtool
libvirt volume storage pool namespace belongs to uvtool, so shouldn't be
messed with by users directly, to prevent future features causing later
collisions. Therefore, I'm dropping this subcommand prior to declaring
the CLI provisionally stable and documenting the remaining subcommands.

We can reimplement it if necessary though. In the meantime, see
http://askubuntu.com/q/299570/7808 for the shell equivalent using virsh
instead.