Correctly handle wait with custom ssh keys
When "uvt-kvm create" is called with --ssh-public-key-file, "uvt-kvm
wait" is no longer able to ssh into the guest by default, because it
can no longer authenticate. This breaks "wait" in this case, when it
would still otherwise be useful to abstract away the remote wait case.

Add options to accomodate these cases:

--without-ssh to explicitly skip the remote wait case, in the case that
the user doesn't want a remote wait, or will arrange the remote wait
himself.

--ssh-private-key-file to specify a private key, in the case that the
user does want a remote wait and can supply a key that can be used.

LP: #1287140